parser/requirements.txt
requests
bs4
pandas
psycopg2-binary